third_party/WebKit/LayoutTests/fast/encoding/api/encoding-labels.html

<!DOCTYPE html>
<title>Encoding API: Encoding labels</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/shared.js"></script>
<script>

encodings_table.forEach(function(section) {
    section.encodings.filter(function(encoding) {
        return encoding.name !== 'replacement';
    }).forEach(function(encoding) {
        var name = encoding.name;
        test(function(){
            encoding.labels.forEach(function(label) {
                assert_equals(new TextDecoder(label).encoding, encoding.name);
            });
        }, 'Labels for: ' + name);
    });
});

</script>
